+++ changelog.md
@@ -5,6 +5,7 @@
 ### 2.6.1, soon...
 
 * fixed: call to undefined function `hash_equals()` on environments with obsolete PHP versions (i.e. < 5.6)
+* fixed: don't run the fixer when a WooCommerce download request is detected
 
 ### 2.6.0, 2018-05-08